requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyYAML==6.0.2
kubernetes==30.1.0
httpx==0.28.1
SQLAlchemy==2.0.38
psycopg[binary]==3.2.9
pytest==8.3.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
